Extract Android SDK Init
